#175 melhorando serializer analise agrotoxico
