+++ plantio/tests/views/test_aplicacacao.py
@@ -156,27 +156,3 @@ class AplicacaoAgrotoxicoAPIViewTest(APIImageTestMixin):
             "Fazer upload de uma imagem válida. O arquivo enviado não é um arquivo de imagem ou está corrompido.",
             response.json()["fotoAgrotoxico"]
         )
-
-    def test_list_aplicacao_produtor(self):
-        recipes.aplicacao.make()
-        aplicacoes = recipes.aplicacao.make(_quantity=2)
-        produtor = aplicacoes[0].plantio.talhao.idPropriedade.produtor
-
-        self.set_client_usuario(produtor.usuario)
-
-        response = self.client.get(self.url)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(response.json()), 2)
-
-    def test_list_aplicacao_tecnico(self):
-        recipes.aplicacao.make()
-        aplicacoes = recipes.aplicacao.make(_quantity=2)
-        tecnico = aplicacoes[0].plantio.talhao.idPropriedade.tecnico
-
-        self.set_client_usuario(tecnico.usuario)
-
-        response = self.client.get(self.url)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(response.json()), 2)
